Separate chat functionality into dedicated Chat component
